pyproject.toml
[tool.poetry]
name = "regex-transformer"
version = "0.1.0"
description = "Small formal languages for training transformers"
authors = ["Robert Cooper <rbcooper@gmail.com>", "Thomas Kwa <kwathomas0@gmail.com>"]
readme = "README.md"
packages = [{include = "regex_transformer"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.1"
automata-lib = "^7.1.0"
frozendict = "^2.3.8"
jupyterlab = "^4.0.2"
transformer-lens = "^1.3.0"
plotly = "^5.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
